Make crash-detection threshold aware of MRES
The stallguard threshold is defined to be 4 phase cycles at 16usteps.
Calculate the threshold using the current ustep resolution, so that it's
consisent[*] also when changing resolution.

[*] tmc2130_st_isr is called for each stepper isr, which in dual/quad
stepping might not correlate with the actual number of performed steps.
But since the threshold is conservative (roughly 1/4 before the actual
step loss in normal circumstances), this should do for the moment.