Recover quicker when SG_DIAG is cleared
When DIAG is cleared the load condition is resolved, meaning either we
stopped moving or the load dropped below the SG threshold.

Instead of clearing blindly "lost usteps" we *should* actually sample
SG_RESULT for a better estimation of recovery. This is expensive
to do in the isr however.

Let's try to be more optimistic and assume transient load spikes rebound
quickly.